refactor: rename jsonld.py to yaml.py in utils (#2818)
