+++ renku/core/migration/m_0009__new_metadata_storage.py
@@ -48,8 +48,8 @@ from renku.core.migration.utils import (
 )
 from renku.core.migration.utils.conversion import convert_dataset
 from renku.core.util import communication
+from renku.core.util.yaml import load_yaml
 from renku.domain_model.entity import Collection, Entity
-from renku.domain_model.jsonld import load_yaml
 from renku.domain_model.project import Project
 from renku.domain_model.provenance.activity import Activity, Association, Generation, Usage
 from renku.domain_model.provenance.agent import Person, SoftwareAgent
